central-dashboard/requirements.txt
flask
requests
pyinstaller
pystray
Pillow
pyobjc-core; sys_platform == 'darwin'
pyobjc-framework-Cocoa; sys_platform == 'darwin'
pyobjc-framework-Quartz; sys_platform == 'darwin'
